Move SeekableAESCipherStream to corresponding sub-package.
